requirements.txt
pyrogram
tgcrypto
configparser
pyautogui
pillow
requests
beautifulsoup4
stegano
cryptography
pillow